backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
python-docx
pydantic>=2.0
python-jose[cryptography]
